Bugfix inadvertent floating point conversion

+++ op25/gr-op25_repeater/apps/rx.py
@@ -418,7 +418,7 @@ class p25_rx_block (gr.top_block):
         e = 0
         if self.last_change_freq > 0:
             err_ppm = round((self.tuning_error*1e6) / float(self.last_change_freq))
-            err_hz = -(self.tuning_error - (err_ppm * (self.last_change_freq / 1e6)))
+            err_hz = -int(self.tuning_error - (err_ppm * (self.last_change_freq / 1e6)))
         if self.options.verbosity >= 10:
             sys.stderr.write('%f frequency_tracking\t%d\t%d\t%d\t%d\t%d\n' % (time.time(), freq_error, self.error_band, self.tuning_error, err_ppm, err_hz))
         if do_freq_update:
